Inventory Update Algorithm (Freecodecamp)
The algorithm adds elements to a two dimensional array where every subarry consists of a string indicating an item of the inventory, and an integer indicating its quantity. The elements are added through another two dimensional array. If the items added are of a type already found in the inventory, the algorithm adds its quantity to that which is already given. If the items added are of a type not yet found in the inventory, the algorithm adds the corresponding subarray at the begginig of the inventory array.